[EXT-2281] Move support links implementation from headers to cpp (#44305)
