+++ ydb/mvp/meta/support_links/ya.make
@@ -2,6 +2,11 @@ LIBRARY()
 
 SRCS(
     entity.cpp
+    grafana_dashboard_common.cpp
+    grafana_dashboard_search_source.cpp
+    grafana_dashboard_source.cpp
+    grafana_logging_source.cpp
+    param_bindings.cpp
     response.cpp
     source.cpp
     support_links_resolver.cpp
